# internal/openapi/swagger.yaml
basePath: /api
...
info:
  contact: {}
  description: HTTP API for the AWG Manager daemon (tunnels, routing, system).
  title: AWG Manager API
  version: "1.0"
paths:
...
  /hook/ndms:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 'Called from router scripts (public). Form fields: type, id, system_name,
        layer, etc.'
      parameters:
      - description: Event type (iflayerchanged, ifcreated, ...)
        in: formData
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: NDMS shell hook
      tags:
      - hook
...
  /tunnels/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Composite tunnels snapshot
      tags:
      - tunnels
...
securityDefinitions:
  CookieAuth:
    description: Session cookie set by POST /auth/login. Omit for public routes.
    in: cookie
    name: awg_session
    type: apiKey
swagger: "2.0"